feat(core): Introduce @sentry/expo-upload-sourcemaps package
Moves the Expo sourcemap upload script into a new scoped workspace
package @sentry/expo-upload-sourcemaps and replaces the original
scripts/expo-upload-sourcemaps.js in @sentry/react-native with a shim
that forwards to it.

The existing sentry-expo-upload-sourcemaps bin in @sentry/react-native
keeps working unchanged, so projects with the bin referenced in
package.json scripts or invoked via npx are not affected.
