+++ .craft.yml
@@ -8,3 +8,5 @@ targets:
     sdks:
       npm:@sentry/react-native:
         includeNames: /^sentry-react-native-\d.*\.tgz$/
+      npm:@sentry/expo-upload-sourcemaps:
+        includeNames: /^sentry-expo-upload-sourcemaps-\d.*\.tgz$/
